requirements.txt
django-admin-autoregister
